Bloody bug fix. Thank you git bisect. illum works
Illumination estimation was failing miserably for some reason. Turns
out that adding a counter broke the `reduce` operation, because the
operand was an array while the right one was a tuple, instead of both
being tuples. For future bug historians, the bug was introduced in
c210b0cecf2309e896c3accc10df9b7b465fffd9, and fixed here.